deletion of old tombstones has been changed to be done at the back of getItemsWithKey:userId:

+++ ADAL/src/cache/ios/ADKeychainTokenCache.m
@@ -285,27 +285,20 @@ static NSString* const s_libraryString = @"MSOpenTech.ADAL." TOSTRING(KEYCHAIN_V
         {
             [itemsKept addObject:item];
         }
-        else
-        {
-            //tombstone will be deleted from cache store if it is too old
-            [self deleteTombstoneIfTooOld:item];
-        }
     }
     SAFE_ARC_AUTORELEASE(itemsKept);
     return itemsKept;
 }
 
-- (void)deleteTombstoneIfTooOld:(ADTokenCacheItem *)item
+//return YES if the tombstone is too old and deleted; NO otherwise.
+- (BOOL)deleteTombstoneIfTooOld:(ADTokenCacheItem *)item
 {
-    if (!item)
-    {
-        return;
-    }
-    
-    if ([[item expiresOn] compare:[NSDate date]] == NSOrderedAscending)
+    if (item && [[item expiresOn] compare:[NSDate date]] == NSOrderedAscending)
     {
         [self deleteItem:item error:nil];
+        return YES;
     }
+    return NO;
 }
 
 - (BOOL)removeAllForClientId:(NSString * __nonnull)clientId
@@ -440,7 +433,8 @@ static NSString* const s_libraryString = @"MSOpenTech.ADAL." TOSTRING(KEYCHAIN_V
     for (NSDictionary* attrs in items)
     {
         ADTokenCacheItem* item = [self itemFromKeyhainAttributes:attrs];
-        if (!item)
+        if (!item ||
+            ([item tombstone] && [self deleteTombstoneIfTooOld:item])) // if item is a tombstone and too old, it is deleted and won't be returned.
         {
             continue;
         }
